requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alabaster==0.7.10         # via sphinx
appnope==0.1.0            # via ipython
attrs==17.4.0             # via pytest
babel==2.5.3              # via sphinx
backcall==0.1.0           # via ipython
bumpversion==0.5.3
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
coverage==4.5.1           # via pytest-cov
decorator==4.3.0          # via ipython, traitlets
django-crum==0.7.2
django-debug-toolbar==1.9.1
django-extensions==2.0.6
django==2.0.4
docutils==0.14            # via sphinx
first==2.0.1              # via pip-tools
flake8==3.5.0
idna==2.6                 # via requests
imagesize==1.0.0          # via sphinx
ipython-genutils==0.2.0   # via traitlets
ipython==6.3.1
jedi==0.12.0              # via ipython
jinja2==2.10              # via sphinx
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==4.1.0     # via pytest
packaging==17.1           # via sphinx
parso==0.2.0              # via jedi
pep8==1.7.1
pexpect==4.5.0            # via ipython
pickleshare==0.7.4        # via ipython
pip-tools==2.0.1
pluggy==0.6.0             # via pytest, tox
prompt-toolkit==1.0.15    # via ipython
ptyprocess==0.5.2         # via pexpect
py==1.5.3                 # via pytest, tox
pycodestyle==2.3.1        # via flake8
pyflakes==1.6.0           # via flake8
pygments==2.2.0           # via ipython, sphinx
pyparsing==2.2.0          # via packaging
pytest-cov==2.5.1
pytest-django==3.2.1
pytest-flake8==1.0.0
pytest-runner==4.2
pytest==3.5.0
pytz==2018.4              # via babel, django
requests==2.18.4          # via sphinx
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via django-extensions, more-itertools, packaging, pip-tools, prompt-toolkit, pytest, sphinx, tox, traitlets
snowballstemmer==1.2.1    # via sphinx
sphinx==1.7.2
sphinxcontrib-websupport==1.0.1  # via sphinx
sqlparse==0.2.4           # via django-debug-toolbar
tox==3.0.0
traitlets==4.3.2          # via ipython
urllib3==1.22             # via requests
virtualenv==15.2.0        # via tox
wcwidth==0.1.7            # via prompt-toolkit
